requirements.txt
fastapi
uvicorn
pandas
requests
sqlalchemy
psycopg2-binary
python-dotenv
tenacity
redis
pydantic
